Fix corruption of PORTD_ODR by main thread LED code.
Fix problem with non-atomic GPIO access in main thread causing debug
pins and radio chip enable to be trashed periodically. The LEDs are
constantly being written to and since the action is not atomic the port
value is set to an old value. On IAR this is fixed by using atomic
instructions (BSET and BRES). On SDCC this is fixed by disabling
interrupts during the LED setting.